Compute knot points of an average shifted histogram

Mainly a utility for the lattice and ggplot2 plotting functions, ash_points() returns the points to be plotted.

ash_points(x, binwidth = NULL, adjust = 1)

Arguments

x
A numeric vector

binwidth
The width of the histogram bins. If NULL (the default) the binwidth will be chosen so that approximately 10 bins cover the data. adjust can be used to to increase or decrease binwidth.

adjust
A number used to scale binwidth.

Value

A data frame containing x and y coordinates of the resulting ASH plot.