return a vector of row or column indices

return a vector of row or column indices

columns(x, default = c())

rows(x, default = c())

Arguments

x
an object that may or may not have any rows or columns

default
what to return if there are no rows or columns

Value

if x has rows or columns, a vector of indices, else default

Examples

dim(iris)

#> [1] 150   5

columns(iris)

#> [1] 1 2 3 4 5

rows(iris)

#>   [1]   1   2   3   4   5   6   7   8   9  10  11  12  13  14  15  16  17  18
#>  [19]  19  20  21  22  23  24  25  26  27  28  29  30  31  32  33  34  35  36
#>  [37]  37  38  39  40  41  42  43  44  45  46  47  48  49  50  51  52  53  54
#>  [55]  55  56  57  58  59  60  61  62  63  64  65  66  67  68  69  70  71  72
#>  [73]  73  74  75  76  77  78  79  80  81  82  83  84  85  86  87  88  89  90
#>  [91]  91  92  93  94  95  96  97  98  99 100 101 102 103 104 105 106 107 108
#> [109] 109 110 111 112 113 114 115 116 117 118 119 120 121 122 123 124 125 126
#> [127] 127 128 129 130 131 132 133 134 135 136 137 138 139 140 141 142 143 144
#> [145] 145 146 147 148 149 150

columns(NULL)

#> NULL

columns("this doesn't have columns")

#> NULL