Convert formulas into standard shapes

These functions convert formulas into standard shapes, including by incorporating a groups argument.

mosaic_formula(
  formula,
  groups = NULL,
  envir = parent.frame(),
  max.slots = 3,
  groups.first = FALSE
)

mosaic_formula_q(
  formula,
  groups = NULL,
  max.slots = 3,
  groups.first = FALSE,
  ...
)

Arguments

formula
a formula

groups
a name used for grouping

envir
the environment in which the resulting formula may be evaluated. May also be NULL, a list, a data frame, or a pairlist.

max.slots
an integer specifying the maximum number of slots for the resulting formula. An error results from trying to create a formula that is too complex.

groups.first
a logical indicating whether groups should be inserted ahead of the condition (else after).

...
additional arguments (currently ignored)

Details

mosaic_formula_q uses nonstandard evaluation of groups that may be necessary for use within other functions. mosaic_formula is a wrapper around mosaic_formula_q and quotes groups before passing it along.

Examples

mosaic_formula( ~ x | z )

#> x ~ z | rlang::enexpr(groups)
#> <environment: 0x7fa0e8fb51c8>

mosaic_formula( ~ x, groups=g )

#> x ~ rlang::enexpr(groups)
#> <environment: 0x7fa0e8fb51c8>

mosaic_formula( y ~ x, groups=g )

#> y ~ x | rlang::enexpr(groups)
#> <environment: 0x7fa0e8fb51c8>

# this is probably not what you want for interactive use.
mosaic_formula_q( y ~ x, groups=g )

#> y ~ x | g
#> <environment: 0x7fa0e8fb51c8>

# but it is for programming
foo <- function(x, groups=NULL) {
    mosaic_formula_q(x, groups=groups, envir=parent.frame())
}
foo( y ~ x , groups = g)

#> y ~ x | groups
#> <environment: 0x7fa0e8fb51c8>