Modified summaries

msummary provides modified summary objects that typically produce output that is either identical to or somewhat terser than their summary() analogs. The contents of the object itself are unchanged (except for an augmented class) so that other downstream functions should work as before.

# S3 method for msummary.lm
print(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)

# S3 method for msummary.glm
print(
  x,
  digits = max(3L, getOption("digits") - 3L),
  symbolic.cor = x$symbolic.cor,
  signif.stars = getOption("show.signif.stars"),
  ...
)

msummary(object, ...)

# S3 method for default
msummary(object, ...)

# S3 method for lm
msummary(object, ...)

# S3 method for glm
msummary(object, ...)

Arguments

x
an object to summarize

digits
desired number of digits to display

symbolic.cor
see summary()

signif.stars
a logical indicating whether to display stars to indicate significance

...
additional arguments

object
an object to summarise

Examples

msummary(lm(Sepal.Length ~ Species, data = iris))

#>                   Estimate Std. Error t value Pr(>|t|)    
#> (Intercept)         5.0060     0.0728  68.762  < 2e-16 ***
#> Speciesversicolor   0.9300     0.1030   9.033 8.77e-16 ***
#> Speciesvirginica    1.5820     0.1030  15.366  < 2e-16 ***
#> 
#> Residual standard error: 0.5148 on 147 degrees of freedom
#> Multiple R-squared:  0.6187,	Adjusted R-squared:  0.6135 
#> F-statistic: 119.3 on 2 and 147 DF,  p-value: < 2.2e-16